add some remnants of rp2040 debugging hell (see message)
This commit has a few commented lines in CMakeLists.txt for compiling
Arduino sketches for the RP2040 using
<https://github.com/earlephilhower/arduino-pico>. I used this to sanity
check if the same use of the Wire library that I used for the Arduino
would work on the RP2040, and to my surprise, it didn't.

Turns out the ATmega328P's TWI interface supports a thing where I2C
masters can be addressed as slaves in a multi-master configuration (see
section 21.9.5 of the datasheet). This is used in the puzzle box, and
not (natively) supported by the RP2040 processor.